rename BlobLoader to PageLoader
