--- packages/loading-indicator/src/BlobLoader/index.ts
@@ -1,3 +1,0 @@
-import BlobLoader from './BlobLoader';
-export { type BlobLoaderProps } from './BlobLoader.types';
-export default BlobLoader;
