+++ packages/loading-indicator/src/PageLoader/PageLoader.spec.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { axe } from 'jest-axe';
 
-import BlobLoader from './BlobLoader';
+import PageLoader from './PageLoader';
 
 describe('packages/loading-indicator/blob-loader', () => {
   describe('a11y', () => {
     test('does not have basic accessibility issues', async () => {
-      const { container } = render(<BlobLoader />);
+      const { container } = render(<PageLoader />);
       const results = await axe(container);
       expect(results).toHaveNoViolations();
     });
@@ -16,7 +16,7 @@ describe('packages/loading-indicator/blob-loader', () => {
     test('renders description', async () => {
       const descriptionText = 'test description';
       const { getByText } = render(
-        <BlobLoader description={descriptionText} />,
+        <PageLoader description={descriptionText} />,
       );
       expect(getByText(descriptionText)).toBeInTheDocument();
     });
