+++ packages/loading-indicator/src/PageLoader/animation.ts
