+++ packages/loading-indicator/src/PageLoader/index.ts
@@ -1,0 +1,3 @@
+import PageLoader from './PageLoader';
+export { type PageLoaderProps } from './PageLoader.types';
+export default PageLoader;
